Nillion package functional
Nillion folder is now correctly setup in the repository and the functions can be called by the agent
