+++ characters/vsa.character.json
@@ -1,7 +1,7 @@
 {
     "name": "VoodooAgent69",
     "clients": [],
-    "modelProvider": "anthropic",
+    "modelProvider": "openai",
     "settings": {
         "voice": {
             "model": "en_GB-danny-low"
