+++ nillion/src/index.d.ts
@@ -1,0 +1,3 @@
+export function storeUserData(data: any): Promise<void>;
+export function retrieveUserData(userId: string): Promise<any>;
+export function testFn(): void;
