--- nillion/storage.d.ts
@@ -1,2 +1,0 @@
-export function storeUserData(data: any): Promise<void>;
-export function retrieveUserData(userId: string): Promise<any>;
