+++ package.json
@@ -84,6 +84,7 @@
   },
   "packageManager": "pnpm@9.15.0",
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "nillion/*"
   ]
 }
