+++ packages/plugin-vsa/package.json
@@ -21,6 +21,7 @@
     "dependencies": {
         "@elizaos/core": "workspace:*",
         "@elizaos/plugin-tee": "workspace:*",
+        "@elizaos/nillion-core": "workspace:*",
         "tsup": "8.3.5"
     },
     "scripts": {
