+++ packages/plugin-vsa/tsconfig.json
@@ -3,7 +3,8 @@
     "compilerOptions": {
         "outDir": "dist",
         "rootDir": "src",
-        "types": ["node"]
+        "types": ["node"],
+        "allowJs": true
     },
     "include": ["src/**/*.ts"]
 }
