+++ pnpm-workspace.yaml
@@ -3,3 +3,4 @@ packages:
     - "packages/*"
     - "client"
     - "agent"
+    - "nillion"
